Updating both tests according to the deprecated and updated OpenAiService classes.
